requirements.txt
crispy-bootstrap5==0.7
dj-database-url==1.0.0
django-configurations==2.4
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
django-registration==3.3
gunicorn==20.1.0
phonenumberslite==8.13.1
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
razorpay==1.3.0
version-lister==0.1.0
whitenoise==6.3.0